Allow absolute paths to compilers
CLion doesn't let you specify a custom PATH for
the CMake executable. (You can specify a custom
path to the CMake executable, but that's not
what we need here.) Therefore, I must qualify
all compiler names as absolute paths.

The ROSBE_BINDIR variable is passed on the CMake
command line, like I18N_LANG and its ilk. If not
specified, it will be ignored. This value must
end with a trailing slash.