Create ignored local.cmake file
My macOS system uses a nonstandard path to bison,
so CMake always selects the system one, which is
too old and incompatible. I know of no way to override
this search path, except by setting BISON_EXECUTABLE
directly.

Since this change is only relevant on my specific
dev box, I have set the variable in local.cmake,
which is a new file included by CMake if it exists,
and that is ignored by Git.