Use nonclient metrics for dialog template if DS_SHELLFONT is set
Previously, to do this you would somehow need to compile
your dialog template with a font size of 0x7FFF. (This
is not an easy task AFAICT, and I don't know if it works
on real Windows.)

docs.microsoft.com states that any dialog template that
sets the DS_SHELLFONT style flag will be laid out using
the current theme message-box font. The docs do not state
that this functionality was removed in Windows Vista,
which my experiments on real Windows have indicated.

One complaint about the real Windows is that its user
interface is so inconsistent compared to Windows 7, and
part of that inconsistency is that some dialogs use the
theme font while others still use Tahoma. Since this is
my fork and I can change it however I please, I am going
to fix this "bug" once and for all here, even if it
makes us not 100% consistent with Real Windows.